agents/requirements.txt
# AWS AgentCore Runtime
bedrock-agentcore

# Strands Agents SDK
strands-agents
strands-agents-tools

# Pydantic for schema validation
pydantic>=2.0.0
